test(macOS): isolate teardown fence coverage
